ver_snap/requirements.txt
altair==5.4.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloup==3.0.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.3
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.4.0
django-filter==24.2
django-jazzmin==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
findspark==2.0.1
Flask==3.0.3
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
glcontext==3.0.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.26.0
isosurfaces==0.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.11.1
moderngl-window==2.4.6
msgpack==1.1.0
multipledispatch==1.0.0
narwhals==1.9.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pycairo==1.27.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pyglet==2.0.17
Pygments==2.18.0
pyparsing==3.1.4
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
rcssmin==1.1.2
referencing==0.35.1
requests==2.32.2
retrying==1.3.4
rich==13.8.1
rjsmin==1.2.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
screeninfo==0.8.1
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
skia-pathops==0.8.0.post1
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
spicy==0.16.0
sqlparse==0.5.1
srt==3.5.3
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.39.0
svgelements==1.9.6
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
